# docs/swagger.yaml
basePath: /cs/v1
definitions:
...
  course-service_internal_domain.Course:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      ownerUserId:
        type: integer
      stats:
        $ref: '#/definitions/course-service_internal_domain.CourseStats'
      status:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
  course-service_internal_domain.CourseSection:
    properties:
      course_id:
        type: integer
      id:
        type: integer
      parent_id:
        type: integer
      sort_order:
        type: integer
      title:
        type: string
    type: object
...
  course-service_internal_domain.CourseStats:
    properties:
      totalStudents:
        type: integer
    type: object
...
  course-service_internal_transport_http_v1_request.CreateCourse:
    properties:
      description:
        type: string
      ownerUserId:
        type: integer
      title:
        type: string
    type: object
  course-service_internal_transport_http_v1_request.CreateCourseSection:
    properties:
      course_id:
        type: integer
      parent_id:
        type: integer
      sort_order:
        type: integer
      title:
        type: string
    type: object
...
  course-service_internal_transport_http_v1_request.UpdateCourse:
    properties:
      description:
        type: string
      ownerUserId:
        type: integer
      status:
        type: integer
      title:
        type: string
    type: object
  course-service_internal_transport_http_v1_request.UpdateCourseSection:
    properties:
      course_id:
        type: integer
      parent_id:
        type: integer
      sort_order:
        type: integer
      title:
        type: string
    type: object
...
  internal_transport_http_v1.ErrorResponse:
    properties:
      message:
        example: bad request
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API for courses, sections, assignments, question banks and tests.
    Tests are also available under `/api` (same handlers as `/cs/v1`).
  title: Course Service API
  version: "1.0"
paths:
...
  /course-sections:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/course-service_internal_domain.CourseSection'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: List course sections
      tags:
      - course-sections
    post:
      consumes:
      - application/json
      parameters:
      - description: Section
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.CreateCourseSection'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/course-service_internal_domain.CourseSection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Create course section
      tags:
      - course-sections
  /course-sections/{id}:
    delete:
      parameters:
      - description: Section ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Delete course section
      tags:
      - course-sections
    get:
      parameters:
      - description: Section ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.CourseSection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Get course section
      tags:
      - course-sections
    put:
      consumes:
      - application/json
      parameters:
      - description: Section ID
        in: path
        name: id
        required: true
        type: integer
      - description: Section
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.UpdateCourseSection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.CourseSection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Update course section
      tags:
      - course-sections
  /courses:
    post:
      consumes:
      - application/json
      parameters:
      - description: Course
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.CreateCourse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/course-service_internal_domain.Course'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Create course
      tags:
      - courses
...
  /courses/{id}:
    delete:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Delete course
      tags:
      - courses
    get:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.Course'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Get course by ID
      tags:
      - courses
    put:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      - description: Course
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.UpdateCourse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.Course'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Update course
      tags:
      - courses
...
schemes:
- http
swagger: "2.0"